veobox: fix potential collision in record ids
The locale date and time was used as ids for
records but it wasn't guaranteeing uniqueness of
the ids as the local time is subject to jet lag.
Consequently it was possible to have the same id
during the time change.
To avoid this the locale date and time is now
suffixed by a unique id.